requirements.txt
python-telegram-bot
yt-dlp
youtube-transcript-api
google-genai